internal: cut deps between assists and diagnostics
