+++ crates/ide_diagnostics/Cargo.toml
@@ -22,7 +22,6 @@ text_edit = { path = "../text_edit", version = "0.0.0" }
 cfg = { path = "../cfg", version = "0.0.0" }
 hir = { path = "../hir", version = "0.0.0" }
 ide_db = { path = "../ide_db", version = "0.0.0" }
-ide_assists = { path = "../ide_assists", version = "0.0.0" }
 
 [dev-dependencies]
 expect-test = "1.1"
