+++ crates/ide_diagnostics/src/missing_fields.rs
@@ -1,7 +1,6 @@
 use either::Either;
 use hir::{db::AstDatabase, InFile};
-use ide_assists::Assist;
-use ide_db::source_change::SourceChange;
+use ide_db::{assists::Assist, source_change::SourceChange};
 use stdx::format_to;
 use syntax::{algo, ast::make, AstNode, SyntaxNodePtr};
 use text_edit::TextEdit;
