+++ crates/ide_diagnostics/src/missing_ok_or_some_in_tail_expr.rs
@@ -1,6 +1,5 @@
 use hir::db::AstDatabase;
-use ide_assists::Assist;
-use ide_db::source_change::SourceChange;
+use ide_db::{assists::Assist, source_change::SourceChange};
 use syntax::AstNode;
 use text_edit::TextEdit;
